requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
boto3==1.40.59
botocore==1.40.59
Brotli==1.1.0
cffi==2.0.0
crispy-bootstrap5==2025.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==4.2.25
django-allauth==65.12.1
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-environ==0.12.0
django-formtools==2.5.1
django-redis==6.0.0
django-storages==1.14.6
fonttools==4.60.1
gunicorn==23.0.0
html5lib==1.1
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
redis==7.0.0
s3transfer==0.14.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.2.3.post1
